Move handleRedirecPromise concurrency test
